# pref: improve marshal & encoding (sec: -39%, B: -70%, allocs: -91%)