--- a/profile/typedef/marshal_test.go
+++ b/profile/typedef/marshal_test.go
@@ -14,6 +14,7 @@ import (
 	"testing"
 
 	"github.com/google/go-cmp/cmp"
+	"github.com/muktihari/fit/kit"
 )
 
 type DefinedBool bool
@@ -33,99 +34,132 @@ type DefinedAny any
 type privateDefinedFloat64 float64
 
 func TestMarshal(t *testing.T) {
-	var i16 int16 = 10
-	var nilptri16 *int16
-
 	tt := []struct {
 		value any
 		err   error
 	}{
-		// // primitive-types
-		{value: false},
-		{value: true},
-		{value: int8(-19)},
-		{value: uint8(129)},
-		{value: int16(1429)},
-		{value: int16(-429)},
-		{value: uint16(9929)},
-		{value: int32(819293429)},
-		{value: int32(-8979123)},
-		{value: uint32(9929)},
-		{value: int64(819293429)},
-		{value: int64(-8979123)},
-		{value: uint64(9929)},
-		{value: float32(819293429.192321)},
-		{value: float32(-8979123.546734)},
-		{value: float64(8192934298908979.192321)},
-		{value: float64(-897912398989898.546734)},
-		{value: "Fit SDK"},
-		{value: ""},
-		{value: []bool{true, false}},
-		{value: []byte{1, 2}},
-		{value: []uint8{1, 2}},
-		{value: []int8{-19}},
-		{value: []uint8{129}},
-		{value: []int16{1429}},
-		{value: []int16{-429}},
-		{value: []uint16{9929}},
-		{value: []int32{819293429}},
-		{value: []int32{-8979123}},
-		{value: []uint32{9929}},
-		{value: []string{"supported"}},
-		{value: []string{""}},
-		{value: []string{string([]byte{'\x00'})}},
-		{value: []int64{819293429}},
-		{value: []int64{-8979123}},
-		{value: []uint64{9929}},
-		{value: []float32{819293429.192321}},
-		{value: []float32{-8979123.546734}},
-		{value: []float64{8192934298908979.192321}},
-		{value: []float64{-897912398989898.546734}},
-		{value: []any{-897912398989898.546734}, err: ErrTypeNotSupported},
-
-		// Defined Types
-		{value: []DefinedInt8{DefinedInt8(1), DefinedInt8(2)}},
-
-		// Types genenerated by fitgen:
-		{value: File(1)},
-		{value: MesgNum(29)},
-		{value: Checksum(1)},
-		{value: FileFlags(1)},
-
-		// User defined type marshaled using reflection:
-		{value: DefinedBool(true)},
-		{value: DefinedInt8(123)},
-		{value: DefinedUint8(123)},
-		{value: DefinedInt16(123)},
-		{value: DefinedUint16(123)},
-		{value: DefinedInt32(123)},
-		{value: DefinedUint32(123)},
-		{value: DefinedInt64(123)},
-		{value: DefinedUint64(123)},
-		{value: DefinedFloat32(123)},
-		{value: DefinedFloat64(123)},
-		{value: DefinedString("Fit SDK")},
-		{value: DefinedString("")},
-
-		// Unsupported types:
+		{value: float32(819293429.192321), err: nil},
 		{value: nil, err: ErrNilValue},
-		{value: complex128(1), err: ErrTypeNotSupported},
-		{value: nilptri16, err: ErrTypeNotSupported},
-		{value: &i16},
-		{value: []int16{10, 10}},
-		{value: []*int16{&i16}},
-		{value: []*int16{nilptri16}, err: ErrTypeNotSupported},
-		{value: []*int16{nil}, err: ErrTypeNotSupported},
-		{value: []DefinedAny{int16(129)}},
-		{value: []DefinedAny{func() {}}, err: ErrTypeNotSupported},
-		{value: []DefinedAny{DefinedAny(nil)}, err: ErrTypeNotSupported},
-		{value: []DefinedAny{[]DefinedAny{}}, err: ErrTypeNotSupported},
-		{value: []privateDefinedFloat64{8.234}},
 	}
 
 	for _, tc := range tt {
 		t.Run(fmt.Sprintf("%T(%v))", tc.value, tc.value), func(t *testing.T) {
 			b, err := Marshal(tc.value, binary.LittleEndian)
+			if !errors.Is(err, tc.err) {
+				t.Fatalf("expected err %s nil, got: %v", tc.err, err)
+			}
+			if err != nil {
+				return
+			}
+			buf := new(bytes.Buffer)
+			err = marshalWithReflectionForTest(buf, tc.value)
+			if err != nil {
+				t.Fatal(err)
+			}
+			if diff := cmp.Diff(b, buf.Bytes()); diff != "" {
+				t.Fatal(diff)
+			}
+		})
+	}
+}
+
+func TestMarshalTo(t *testing.T) {
+	var i16 int16 = 10
+	var nilptri16 *int16
+
+	tt := []struct {
+		b     *[]byte
+		value any
+		err   error
+	}{
+		// // primitive-types
+		{b: kit.Ptr([]byte{}), value: false},
+		{b: kit.Ptr([]byte{}), value: true},
+		{b: kit.Ptr([]byte{}), value: int8(-19)},
+		{b: kit.Ptr([]byte{}), value: uint8(129)},
+		{b: kit.Ptr([]byte{}), value: int16(1429)},
+		{b: kit.Ptr([]byte{}), value: int16(-429)},
+		{b: kit.Ptr([]byte{}), value: uint16(9929)},
+		{b: kit.Ptr([]byte{}), value: int32(819293429)},
+		{b: kit.Ptr([]byte{}), value: int32(-8979123)},
+		{b: kit.Ptr([]byte{}), value: uint32(9929)},
+		{b: kit.Ptr([]byte{}), value: int64(819293429)},
+		{b: kit.Ptr([]byte{}), value: int64(-8979123)},
+		{b: kit.Ptr([]byte{}), value: uint64(9929)},
+		{b: kit.Ptr([]byte{}), value: float32(819293429.192321)},
+		{b: kit.Ptr([]byte{}), value: float32(-8979123.546734)},
+		{b: kit.Ptr([]byte{}), value: float64(8192934298908979.192321)},
+		{b: kit.Ptr([]byte{}), value: float64(-897912398989898.546734)},
+		{b: kit.Ptr([]byte{}), value: "Fit SDK"},
+		{b: kit.Ptr([]byte{}), value: ""},
+		{b: kit.Ptr([]byte{}), value: []bool{true, false}},
+		{b: kit.Ptr([]byte{}), value: []byte{1, 2}},
+		{b: kit.Ptr([]byte{}), value: []uint8{1, 2}},
+		{b: kit.Ptr([]byte{}), value: []int8{-19}},
+		{b: kit.Ptr([]byte{}), value: []uint8{129}},
+		{b: kit.Ptr([]byte{}), value: []int16{1429}},
+		{b: kit.Ptr([]byte{}), value: []int16{-429}},
+		{b: kit.Ptr([]byte{}), value: []uint16{9929}},
+		{b: kit.Ptr([]byte{}), value: []int32{819293429}},
+		{b: kit.Ptr([]byte{}), value: []int32{-8979123}},
+		{b: kit.Ptr([]byte{}), value: []uint32{9929}},
+		{b: kit.Ptr([]byte{}), value: []string{"supported"}},
+		{b: kit.Ptr([]byte{}), value: []string{""}},
+		{b: kit.Ptr([]byte{}), value: []string{string([]byte{'\x00'})}},
+		{b: kit.Ptr([]byte{}), value: []int64{819293429}},
+		{b: kit.Ptr([]byte{}), value: []int64{-8979123}},
+		{b: kit.Ptr([]byte{}), value: []uint64{9929}},
+		{b: kit.Ptr([]byte{}), value: []float32{819293429.192321}},
+		{b: kit.Ptr([]byte{}), value: []float32{-8979123.546734}},
+		{b: kit.Ptr([]byte{}), value: []float64{8192934298908979.192321}},
+		{b: kit.Ptr([]byte{}), value: []float64{-897912398989898.546734}},
+		{b: kit.Ptr([]byte{}), value: []any{-897912398989898.546734}, err: ErrTypeNotSupported},
+
+		// Defined Types
+		{b: kit.Ptr([]byte{}), value: []DefinedInt8{DefinedInt8(1), DefinedInt8(2)}},
+
+		// Types genenerated by fitgen:
+		{b: kit.Ptr([]byte{}), value: File(1)},
+		{b: kit.Ptr([]byte{}), value: MesgNum(29)},
+		{b: kit.Ptr([]byte{}), value: Checksum(1)},
+		{b: kit.Ptr([]byte{}), value: FileFlags(1)},
+
+		// User defined type marshaled using reflection:
+		{b: kit.Ptr([]byte{}), value: DefinedBool(true)},
+		{b: kit.Ptr([]byte{}), value: DefinedInt8(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedUint8(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedInt16(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedUint16(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedInt32(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedUint32(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedInt64(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedUint64(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedFloat32(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedFloat64(123)},
+		{b: kit.Ptr([]byte{}), value: DefinedString("Fit SDK")},
+		{b: kit.Ptr([]byte{}), value: DefinedString("")},
+
+		// Unsupported types:
+		{b: kit.Ptr([]byte{}), value: nil, err: ErrNilValue},
+		{b: kit.Ptr([]byte{}), value: complex128(1), err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: nilptri16, err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: &i16},
+		{b: kit.Ptr([]byte{}), value: []int16{10, 10}},
+		{b: kit.Ptr([]byte{}), value: []*int16{&i16}},
+		{b: kit.Ptr([]byte{}), value: []*int16{nilptri16}, err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: []*int16{nil}, err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: []DefinedAny{int16(129)}},
+		{b: kit.Ptr([]byte{}), value: []DefinedAny{func() {}}, err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: []DefinedAny{DefinedAny(nil)}, err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: []DefinedAny{[]DefinedAny{}}, err: ErrTypeNotSupported},
+		{b: kit.Ptr([]byte{}), value: []privateDefinedFloat64{8.234}},
+
+		{b: nil, value: uint8(129), err: ErrNilDest},
+	}
+
+	for _, tc := range tt {
+		t.Run(fmt.Sprintf("%T(%v))", tc.value, tc.value), func(t *testing.T) {
+			err := MarshalTo(tc.b, tc.value, binary.LittleEndian)
 			if !errors.Is(err, tc.err) {
 				t.Fatalf("expected err: %v, got: %v", tc.err, err)
 			}
@@ -138,12 +172,12 @@ func TestMarshal(t *testing.T) {
 				t.Fatal(err)
 			}
 
-			if len(b) == 0 && len(buf.Bytes()) == 0 {
+			if len(*tc.b) == 0 && len(buf.Bytes()) == 0 {
 				return
 			}
 
-			if diff := cmp.Diff(b, buf.Bytes()); diff != "" {
-				fmt.Printf("value: %v, b: %v, buf: %v\n", tc.value, b, buf.Bytes())
+			if diff := cmp.Diff(*tc.b, buf.Bytes()); diff != "" {
+				fmt.Printf("value: %v, b: %v, buf: %v\n", tc.value, *tc.b, buf.Bytes())
 				t.Fatal(diff)
 			}