… profile/typedef/marshal.go