benchstat:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/encoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
         │    old.txt    │               new.txt                │
         │    sec/op     │    sec/op     vs base                │
Encode-4   152.86m ± 33%   92.27m ± 26%  -39.64% (p=0.001 n=10)

         │    old.txt    │               new.txt                │
         │     B/op      │     B/op      vs base                │
Encode-4   10.301Mi ± 0%   3.074Mi ± 0%  -70.16% (p=0.000 n=10)

         │   old.txt    │               new.txt               │
         │  allocs/op   │  allocs/op   vs base                │
Encode-4   1200.1k ± 0%   100.0k ± 0%  -91.66% (p=0.000 n=10)

```